Changed Width and changed JS
I thought the previous JS of changing the input field color would work when I would hover over the fields but it did not, it should work now when my mouse hovers over the field. I also changed the width of the inputs on mobile screens so it's more touch friendly.